Replaced size_t with int.

+++ mp_core.c
@@ -76,7 +76,7 @@ static struct {
     int visible;                /* offset to the first visible character */
     int cursor;                 /* offset to cursor */
     wchar_t *ptr;               /* pointer to joined data */
-    size_t size;                /* size of joined data */
+    int size;                   /* size of joined data */
     int matchparen_offset;      /* offset to matched paren */
     int matchparen_o_attr;      /* original attribute there */
     int cursor_o_attr;          /* original attribute under cursor */
